Swtich to Vite because I like it better.
