+++ src/App.css
@@ -1,5 +1,12 @@
+#root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
+}
+
 * {
- box-sizing: border-box;
+  box-sizing: border-box;
 }
 
 fieldset, textarea {
