+++ src/components/CandidateInfo/CandidateInfo.jsx
