+++ vite.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from "vite";
+
+import react from "@vitejs/plugin-react";
+import mkcert from "vite-plugin-mkcert";
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react(), mkcert()],
+  server: {
+    https: true,
+  },
+});
